+++ src/views/payment-sheet-views/card-view.js
@@ -23,19 +23,19 @@ CardView.ID = CardView.prototype.ID = constants.paymentOptionIDs.card;
 
 CardView.prototype.initialize = function () {
   var cvvFieldGroup, postalCodeFieldGroup, hfOptions;
-  var cardholderNameField = this.getElementById('cardholder-name-field-group');
+  var cardholderNameGroup = this.getElementById('cardholder-name-field-group');
   var cardIcons = this.getElementById('card-view-icons');
 
   this.merchantConfiguration = this.model.merchantConfiguration.card || {};
   this.merchantConfiguration.vault = this.merchantConfiguration.vault || {};
+  this.hasCardholderName = Boolean(this.merchantConfiguration.cardholderName);
+  this.cardholderNameRequired = this.hasCardholderName && this.merchantConfiguration.cardholderName.required === true;
   hfOptions = this._generateHostedFieldsOptions();
 
   cardIcons.innerHTML = cardIconHTML;
   this._hideUnsupportedCardIcons();
 
   this.hasCVV = hfOptions.fields.cvv;
-  this.hasCardholderName = Boolean(this.merchantConfiguration.cardholderName);
-  this.cardholderNameInput = cardholderNameField.querySelector('input');
   this.saveCardInput = this.getElementById('save-card-input');
   this.cardNumberIcon = this.getElementById('card-number-icon');
   this.cardNumberIconSvg = this.getElementById('card-number-icon-svg');
@@ -43,22 +43,10 @@ CardView.prototype.initialize = function () {
   this.cvvIconSvg = this.getElementById('cvv-icon-svg');
   this.cvvLabelDescriptor = this.getElementById('cvv-label-descriptor');
   this.fieldErrors = {};
-  this.extraInputs = [
-    {
-      fieldName: 'cardholderName',
-      enabled: this.hasCardholderName,
-      required: this.hasCardholderName && this.merchantConfiguration.cardholderName.required,
-      requiredError: this.strings.fieldEmptyForCardholderName,
-      validations: [
-        {
-          isValid: function (value) {
-            return value.length < 256;
-          },
-          error: this.strings.fieldTooLongForCardholderName
-        }
-      ]
-    }
-  ];
+
+  if (!this.hasCardholderName) {
+    cardholderNameGroup.parentNode.removeChild(cardholderNameGroup);
+  }
 
   if (!this.hasCVV) {
     cvvFieldGroup = this.getElementById('cvv-field-group');
@@ -69,14 +57,6 @@ CardView.prototype.initialize = function () {
     postalCodeFieldGroup = this.getElementById('postal-code-field-group');
     postalCodeFieldGroup.parentNode.removeChild(postalCodeFieldGroup);
   }
-
-  this.extraInputs.forEach(function (extraInput) {
-    if (extraInput.enabled) {
-      this._setupExtraInput(extraInput);
-    } else {
-      this._removeExtraInput(extraInput);
-    }
-  }.bind(this));
 
   if (!this.model.isGuestCheckout && this.merchantConfiguration.vault.allowVaultCardOverride === true) {
     classList.remove(this.getElementById('save-card-field-group'), 'braintree-hidden');
@@ -105,45 +85,6 @@ CardView.prototype.initialize = function () {
   }.bind(this));
 };
 
-CardView.prototype._setupExtraInput = function (extraInput) {
-  var self = this;
-  var fieldNameKebab = camelCaseToKebabCase(extraInput.fieldName);
-  var field = this.getElementById(fieldNameKebab + '-field-group');
-  var input = field.querySelector('input');
-  var nameContainer = field.querySelector('.braintree-form__hosted-field');
-
-  input.addEventListener('keyup', function () {
-    var valid = self._validateExtraInput(extraInput, true);
-
-    classList.toggle(nameContainer, 'braintree-form__field--valid', valid);
-
-    if (valid) {
-      self.hideFieldError(extraInput.fieldName);
-    }
-
-    self._sendRequestableEvent();
-  }, false);
-
-  if (extraInput.required) {
-    input.addEventListener('blur', function () {
-      // the active element inside the blur event is the document.body
-      // by taking it out of the event loop, we can detect the new
-      // active element (hosted field or other card view element)
-      setTimeout(function () {
-        if (isCardViewElement()) {
-          self._validateExtraInput(extraInput, true);
-        }
-      }, 0);
-    }, false);
-  }
-};
-
-CardView.prototype._removeExtraInput = function (extraInput) {
-  var field = this.getElementById(camelCaseToKebabCase(extraInput.fieldName) + '-field-group');
-
-  field.parentNode.removeChild(field);
-};
-
 CardView.prototype._sendRequestableEvent = function () {
   if (!this._isTokenizing) {
     this.model.setPaymentMethodRequestable({
@@ -161,6 +102,10 @@ CardView.prototype._generateHostedFieldsOptions = function () {
   var options = {
     client: this.client,
     fields: {
+      cardholderName: {
+        container: this._getFieldContainer('cardholder-name'),
+        placeholder: this.strings.cardholderNamePlaceholder
+      },
       number: {
         container: this._getFieldContainer('number'),
         placeholder: generateCardNumberPlaceholder()
@@ -203,6 +148,10 @@ CardView.prototype._generateHostedFieldsOptions = function () {
       }
     }
   };
+
+  if (!this.hasCardholderName) {
+    delete options.fields.cardholderName;
+  }
 
   if (!hasCVVChallenge) {
     delete options.fields.cvv;
@@ -280,10 +229,14 @@ CardView.prototype._validateForm = function (showFieldErrors) {
     }
 
     if (field.isEmpty) {
-      isValid = false;
+      if (!this.cardholderNameRequired && key === 'cardholderName') {
+        isValid = true;
+      } else {
+        isValid = false;
 
-      if (showFieldErrors) {
-        this.showFieldError(key, this.strings['fieldEmptyFor' + capitalize(key)]);
+        if (showFieldErrors) {
+          this.showFieldError(key, this.strings['fieldEmptyFor' + capitalize(key)]);
+        }
       }
     } else if (!field.isValid) {
       isValid = false;
@@ -308,48 +261,7 @@ CardView.prototype._validateForm = function (showFieldErrors) {
     }
   }
 
-  if (this.extraInputs) {
-    this.extraInputs.forEach(function (extraInput) {
-      var fieldIsValid;
-
-      if (!extraInput.enabled) {
-        return;
-      }
-
-      fieldIsValid = this._validateExtraInput(extraInput, showFieldErrors);
-
-      isValid = isValid && fieldIsValid;
-    }.bind(this));
-  }
-
   return isValid;
-};
-
-CardView.prototype._validateExtraInput = function (extraInput, showFieldError) {
-  var fieldNameKebab = camelCaseToKebabCase(extraInput.fieldName);
-  var field = this.getElementById(fieldNameKebab + '-field-group');
-  var input = field.querySelector('input');
-  var valid = true;
-
-  if (extraInput.required) {
-    valid = input.value.length > 0;
-
-    if (!valid && showFieldError) {
-      this.showFieldError(extraInput.fieldName, extraInput.requiredError);
-    }
-  }
-
-  extraInput.validations.forEach(function (validation) {
-    var validationPassed = validation.isValid(input.value);
-
-    if (!validationPassed && showFieldError) {
-      this.showFieldError(extraInput.fieldName, validation.error);
-    }
-
-    valid = valid && validationPassed;
-  }.bind(this));
-
-  return valid;
 };
 
 CardView.prototype.getPaymentMethod = function () { // eslint-disable-line consistent-return
@@ -379,8 +291,10 @@ CardView.prototype.tokenize = function () {
     return Promise.reject(new DropinError(constants.errors.NO_PAYMENT_METHOD_ERROR));
   }
 
-  if (this.hasCardholderName) {
-    tokenizeOptions.cardholderName = this.cardholderNameInput.value;
+  if (!this.cardholderNameRequired && state.fields.cardholderName && state.fields.cardholderName.isEmpty) {
+    tokenizeOptions.fieldsToTokenize = Object.keys(state.fields).filter(function (field) {
+      return field !== 'cardholderName';
+    });
   }
 
   self._isTokenizing = true;
@@ -392,10 +306,6 @@ CardView.prototype.tokenize = function () {
       Object.keys(state.fields).forEach(function (field) {
         self.hostedFieldsInstance.clear(field);
       });
-
-      if (self.hasCardholderName) {
-        self.cardholderNameInput.value = '';
-      }
     }
 
     if (self._shouldVault()) {
@@ -452,7 +362,7 @@ CardView.prototype.showFieldError = function (field, errorMessage) {
   fieldError = this.fieldErrors[field];
   fieldError.innerHTML = errorMessage;
 
-  if (input && isNormalFieldElement(input)) {
+  if (input) {
     input.setAttribute('aria-invalid', true);
   } else {
     this.hostedFieldsInstance.setAttribute({
@@ -477,7 +387,7 @@ CardView.prototype.hideFieldError = function (field) {
 
   classList.remove(fieldGroup, 'braintree-form__field-group--has-error');
 
-  if (input && isNormalFieldElement(input)) {
+  if (input) {
     input.removeAttribute('aria-invalid');
   } else {
     this.hostedFieldsInstance.removeAttribute({
@@ -515,7 +425,7 @@ CardView.prototype._onBlurEvent = function (event) {
 
   classList.remove(fieldGroup, 'braintree-form__field-group--is-focused');
 
-  if (shouldApplyFieldEmptyError(field)) {
+  if (this._shouldApplyFieldEmptyError(event.emittedBy, field)) {
     this.showFieldError(event.emittedBy, this.strings['fieldEmptyFor' + capitalize(event.emittedBy)]);
   } else if (!field.isEmpty && !field.isValid) {
     this.showFieldError(event.emittedBy, this.strings['fieldInvalidFor' + capitalize(event.emittedBy)]);
@@ -527,7 +437,7 @@ CardView.prototype._onBlurEvent = function (event) {
     // when focusing on a field by clicking the label,
     // we need to wait a bit for the iframe to be
     // focused properly before applying validations
-    if (shouldApplyFieldEmptyError(field)) {
+    if (this._shouldApplyFieldEmptyError(event.emittedBy, field)) {
       this.showFieldError(event.emittedBy, this.strings['fieldEmptyFor' + capitalize(event.emittedBy)]);
     }
   }.bind(this), 150);
@@ -611,13 +521,13 @@ CardView.prototype.onSelection = function () {
     return;
   }
 
-  if (this.hasCardholderName) {
-    setTimeout(function () { // wait until input is visible
-      this.cardholderNameInput.focus();
-    }.bind(this), 1);
-  } else {
-    this.hostedFieldsInstance.focus('number');
-  }
+  setTimeout(function () {
+    if (this.hasCardholderName) {
+      this.hostedFieldsInstance.focus('cardholderName');
+    } else {
+      this.hostedFieldsInstance.focus('number');
+    }
+  }.bind(this), 50);
 };
 
 CardView.prototype._hideUnsupportedCardIcons = function () {
@@ -648,19 +558,23 @@ CardView.isEnabled = function (options) {
   return Promise.resolve(!disabledByMerchant && gatewayConfiguration.creditCards.supportedCardTypes.length > 0);
 };
 
-function isNormalFieldElement(element) {
-  return element.id.indexOf('braintree__card-view-input') !== -1;
-}
+CardView.prototype._shouldApplyFieldEmptyError = function (fieldId, field) {
+  if (!field.isEmpty) {
+    return false;
+  }
 
-function shouldApplyFieldEmptyError(field) {
-  return field.isEmpty && isCardViewElement();
-}
+  if (fieldId === 'cardholderName' && !this.cardholderNameRequired) {
+    return false;
+  }
+
+  return isCardViewElement();
+};
 
 function isCardViewElement() {
   var activeId = document.activeElement && document.activeElement.id;
   var isHostedFieldsElement = document.activeElement instanceof HTMLIFrameElement && activeId.indexOf('braintree-hosted-field') !== -1;
 
-  return isHostedFieldsElement || isNormalFieldElement(document.activeElement);
+  return isHostedFieldsElement;
 }
 
 function camelCaseToKebabCase(string) {